backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0

# Database and ORM
sqlmodel==0.0.14
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# HTTP client
httpx==0.26.0
requests==2.31.0

# Monitoring and observability
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Security and authentication
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.9

# OpenAI integration
openai==1.12.0

# Symbolic reasoning engines
z3-solver==4.12.2.0
pyswip==0.3.3
networkx==3.2.1

# Data processing
pandas==2.2.0
numpy==1.26.4
PyYAML==6.0.1

# Graph visualization
matplotlib==3.8.2
plotly==5.17.0
graphviz==0.20.1

# Utilities
python-dotenv==1.0.1
psutil==5.9.8

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
